+++ ouzel/input/macos/GamepadDeviceMacOS.hpp
@@ -14,7 +14,7 @@ namespace ouzel
         {
         public:
             GamepadDeviceMacOS(InputSystem& initInputSystem,
-                               uint32_t initId):
+                               uintptr_t initId):
                 GamepadDevice(initInputSystem, initId)
             {
             }
