+++ ouzel/input/macos/MouseDeviceMacOS.hpp
@@ -16,7 +16,7 @@ namespace ouzel
         {
         public:
             MouseDeviceMacOS(InputSystem& initInputSystem,
-                             uint32_t initId):
+                             uintptr_t initId):
                 MouseDevice(initInputSystem, initId)
             {
             }
